requirements.txt
tensorflow-gpu==1.15  # GPU version of TensorFlow
tqdm
six